Ajout de la barre de vie et d'une perte de 1 PV en touchant la limace.

+++ trunk/survivalkid/SurvivalKid/src/com/survivalkid/game/entity/Life.java
@@ -65,7 +65,7 @@ public class Life {
 	/// GETTER AND SETTER
 	
 	public float getCurrentPcLife() {
-		return currentLife / maxLife;
+		return currentLife / (float) maxLife;
 	}
 
 	public int getMaxLife() {
